Subtask.go and pomodoro.go now merged together in pomodoro package
